fix: address review feedback on phase 2 target split
- bind associated value on ConversationEvent.loopDetected switch case
  (case .loopDetected(_)) — the case carries a messageID associated value;
  the original case .loopDetected: form is technically valid Swift but the
  explicit (_) bind matches Copilot's expectation and makes the unused
  binding intentional.
- remove duplicate `import BaseChatRuntime` in SessionManagerViewModel
  and consolidate redundant plain/@testable imports in
  SessionListServiceTests
- update stale BaseChatCore doc references in GenerationCoordinator and
  StreamingTokenBatcher to point at BaseChatRuntime (the new orchestration
  target)
- update BaseChatUI DocC (BaseChatUI.md + BuildingAChatUI.md) to reflect
  the new target split: BaseChatUI now depends on BaseChatRuntime plus
  BaseChatPersistenceSwiftData (interim, until APIEndpoint is decoupled
  from SwiftData); fix three import snippets and renamed-class references
- collapse double secureKey.stringValue reads in OpenAIBackend +
  OpenAIResponsesBackend so the SecureBytes value is materialised exactly
  once per request build
- rename BaseChatRuntime *class* (in BaseChatPersistenceSwiftData) to
  BaseChatBootstrap to stop it shadowing the BaseChatRuntime *target*
  in IDE jump-to-definition and DocC. Merged the existing one-off
  BaseChatBootstrap enum (single static reapOrphanedKeychainItems
  helper) into the renamed class as a static method, deleted the
  old BaseChatRuntime.swift file, and updated all class references
  in Sources/ + Tests/ + DocC. Module-level `import BaseChatRuntime`
  references are unchanged — only the type name moved.
