+++ Sources/BaseChatUI/BaseChatUI.docc/BaseChatUI.md
@@ -4,14 +4,15 @@ SwiftUI views and view models for building on-device and cloud-connected chat in
 
 ## Overview
 
-BaseChatUI provides the view layer for BaseChatKit. It depends only on ``BaseChatCore`` — it has no knowledge of specific inference backends. Drop ``ChatView`` into your app and supply a ``ChatViewModel`` to get a fully-featured chat interface: streaming generation, model selection, and session management.
+BaseChatUI provides the view layer for BaseChatKit. It depends on ``BaseChatRuntime`` (the persistence-free orchestration target) and, as an interim measure, on ``BaseChatPersistenceSwiftData`` for the `APIEndpoint` `@Model` type — that coupling is scheduled to be removed once the endpoint surface is decoupled from SwiftData. It has no knowledge of specific inference backends. Drop ``ChatView`` into your app and supply a ``ChatViewModel`` to get a fully-featured chat interface: streaming generation, model selection, and session management.
 
 `ChatInputBar` automatically exposes image attachments only when the active backend's ``BackendCapabilities/supportsVision`` flag is `true`. If a host routes image-bearing history to a text-only backend anyway, `GenerationCoordinator` fails fast rather than silently flattening the images away.
 
 ### Minimum wiring
 
 ```swift
-import BaseChatCore
+import BaseChatRuntime
+import BaseChatPersistenceSwiftData
 import BaseChatInference
 import BaseChatBackends
 import BaseChatUI
@@ -20,11 +21,11 @@ import SwiftUI
 
 @main
 struct MyApp: App {
-    let runtime: BaseChatRuntime
+    let runtime: BaseChatBootstrap
     let chatVM: ChatViewModel
 
     init() {
-        let runtime = try! BaseChatRuntime(
+        let runtime = try! BaseChatBootstrap(
             configuration: BaseChatConfiguration(
                 appName: "MyApp",
                 bundleIdentifier: Bundle.main.bundleIdentifier ?? "com.example.myapp"
