+++ Tests/BaseChatCoreTests/RuntimeBootstrapMilestoneTests.swift
@@ -10,7 +10,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let originalConfiguration = BaseChatConfiguration.shared
         defer { BaseChatConfiguration.shared = originalConfiguration }
 
-        let (progress, task) = BaseChatRuntime.build(
+        let (progress, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Milestone Order",
                 bundleIdentifier: "com.basechatkit.milestone-tests.order.\(UUID().uuidString)"
@@ -31,7 +31,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let originalConfiguration = BaseChatConfiguration.shared
         defer { BaseChatConfiguration.shared = originalConfiguration }
 
-        let (progress, task) = BaseChatRuntime.build(
+        let (progress, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Complete Is Last",
                 bundleIdentifier: "com.basechatkit.milestone-tests.last.\(UUID().uuidString)"
@@ -52,7 +52,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let originalConfiguration = BaseChatConfiguration.shared
         defer { BaseChatConfiguration.shared = originalConfiguration }
 
-        let (progress, task) = BaseChatRuntime.build(
+        let (progress, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Failure Stream",
                 bundleIdentifier: "com.basechatkit.milestone-tests.failure.\(UUID().uuidString)"
@@ -79,7 +79,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let originalConfiguration = BaseChatConfiguration.shared
         defer { BaseChatConfiguration.shared = originalConfiguration }
 
-        let (_, task) = BaseChatRuntime.build(
+        let (_, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Rollback",
                 bundleIdentifier: "com.basechatkit.milestone-tests.rollback.\(UUID().uuidString)"
@@ -99,7 +99,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let injectedService = InferenceService()
         let prebuiltContainer = try ModelContainerFactory.makeInMemoryContainer()
 
-        let (progress, task) = BaseChatRuntime.build(
+        let (progress, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Wiring Identity",
                 bundleIdentifier: "com.basechatkit.milestone-tests.wiring.\(UUID().uuidString)"
@@ -134,7 +134,7 @@ final class RuntimeBootstrapMilestoneTests: XCTestCase {
         let originalConfiguration = BaseChatConfiguration.shared
         defer { BaseChatConfiguration.shared = originalConfiguration }
 
-        let (progress, task) = BaseChatRuntime.build(
+        let (progress, task) = BaseChatBootstrap.build(
             configuration: BaseChatConfiguration(
                 appName: "Build Persistence",
                 bundleIdentifier: "com.basechatkit.milestone-tests.persistence.\(UUID().uuidString)"
