+++ Tests/BaseChatUITests/RuntimeConfigurationTests.swift
@@ -18,7 +18,7 @@ final class RuntimeConfigurationTests: XCTestCase {
         }
         defer { defaults.removePersistentDomain(forName: suiteName) }
 
-        let runtime = try BaseChatRuntime(
+        let runtime = try BaseChatBootstrap(
             configuration: BaseChatConfiguration(
                 appName: "Runtime UI Tests",
                 bundleIdentifier: "com.basechatkit.runtime-ui-tests"
@@ -69,7 +69,7 @@ final class RuntimeConfigurationTests: XCTestCase {
         }
         defer { defaults.removePersistentDomain(forName: suiteName) }
 
-        let runtime = try BaseChatRuntime(
+        let runtime = try BaseChatBootstrap(
             configuration: BaseChatConfiguration(
                 appName: "Bootstrap Tests",
                 bundleIdentifier: "com.basechatkit.runtime-ui-tests.bootstrap"
